hub/commands: ProcessTransfer - move aux functions to helper.h and assert against actual error string
